requirements.txt
# Crawling
crawl4ai>=0.4.0
beautifulsoup4>=4.12.0
markdownify>=0.13.1

# Azure OpenAI
openai>=1.30.0

# Pinecone
pinecone>=5.0.0

# Redis
redis>=5.0.0

# Deepgram
deepgram-sdk>=3.0.0

# Audio
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0

# OpenAI standard (for Whisper + TTS, separate from Azure)
openai>=1.30.0   # already present

# ElevenLabs
elevenlabs>=1.0.0
httpx>=0.27.0

faiss-cpu

pydub>=0.25.1